refactor: remove redundent codes
