+++ src/etha/comm/__init__.py
@@ -1,11 +1,10 @@
 """Communication utilities."""
 
 from .get_p2p_map import get_p2p_map
-from .comm_methods import p2p_communicate, get_shard_tensor_shape, gather_broadcast_communicate
+from .comm_methods import p2p_communicate, gather_broadcast_communicate
 
 __all__ = [
     "get_p2p_map",
     "p2p_communicate",
     "gather_broadcast_communicate",
-    "get_shard_tensor_shape",
 ]
