+++ src/etha/comm/comm_methods.py
@@ -4,7 +4,7 @@ import logging
 
 import torch
 import torch.distributed as dist
-from torch.distributed._tensor import Shard, DTensor, DeviceMesh, distribute_tensor
+from torch.distributed._tensor import DTensor, DeviceMesh, distribute_tensor
 from torch.distributed.tensor.placement_types import Placement
 
 from .comm_execution import execute_naive, prepare_recv_buffers, prepare_send_buffers
@@ -44,74 +44,6 @@ def gather_broadcast_communicate(
         received_tensor = distribute_tensor(gathered_tensor, target_mesh, target_specs)
 
     return received_tensor
-
-
-def get_shard_shape(device_mesh: tuple[int, ...], placements: tuple[Placement, ...], tensor_ndim: int) -> list[int]:
-    """Calculate shard shape from device mesh and placements."""
-    shard_shape = [1] * tensor_ndim
-    for i, placement in enumerate(placements):
-        if isinstance(placement, Shard):
-            shard_shape[placement.dim] *= device_mesh[i]
-    return shard_shape
-
-
-def get_shard_tensor_shape(
-    origin_full_shape: torch.Size, target_device_mesh: DeviceMesh, placements: tuple[Placement, ...]
-) -> torch.Size:
-    target_shard_shape = list(origin_full_shape)
-    for i, placement in enumerate(placements):
-        if isinstance(placement, Shard):
-            mesh_dim_size = target_device_mesh.mesh.shape[i]
-            if target_shard_shape[placement.dim] % mesh_dim_size != 0:
-                raise ValueError(
-                    f"Dimension {placement.dim} of tensor with shape "
-                    f"{target_shard_shape[placement.dim]} is not divisible by "
-                    f"mesh dimension {i} with size {mesh_dim_size}"
-                )
-            target_shard_shape[placement.dim] //= mesh_dim_size
-    return torch.Size(target_shard_shape)
-
-
-def assemble_target_tensor(
-    rank: int,
-    reverse_map: dict[int, dict[tuple, list[tuple[int, tuple]]]],
-    received_data: dict[tuple[int, tuple], torch.Tensor],
-    recv_buffers: dict[tuple[tuple, int, tuple], torch.Tensor],
-    target_num_slicers: list[int],
-    final_tensor: torch.Tensor,
-) -> torch.Tensor | None:
-    """Assemble final tensor from received chunks.
-
-    Returns:
-        Final assembled tensor if rank is a receiver, None otherwise.
-    """
-    if rank not in reverse_map:
-        return None
-
-    for target_idx, src_list in reverse_map[rank].items():
-        for src_rank, sidx in src_list:
-            if src_rank == rank:
-                if (src_rank, sidx) not in received_data:
-                    continue
-                buf = received_data[(src_rank, sidx)]
-            else:
-                buf_key = (target_idx, src_rank, sidx)
-                buf = recv_buffers[buf_key]
-
-            slice_ranges = []
-            for dim, coord in enumerate(target_idx):
-                if target_num_slicers[dim] > 1:
-                    slice_size = final_tensor.shape[dim] // target_num_slicers[dim]
-                    start = coord * slice_size
-                    end = start + slice_size
-                    slice_ranges.append(slice(start, end))
-                else:
-                    slice_ranges.append(slice(None))
-
-            final_tensor[tuple(slice_ranges)].copy_(buf)
-
-            if src_rank != rank:
-                del recv_buffers[buf_key]
 
 
 def p2p_communicate(
